+++ src/routes/index.svelte
@@ -1,5 +1,5 @@
 <script>
-  import IconButton from "../components/IconButton.svelte";
+  import IconButton from "../components/Elements/IconButton.svelte";
 
   const versionNumber = "v0.9 ©2019";
 </script>
